requirements.txt
streamlit==1.39.0
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
